Add platform defop test, force renames in copy tests, and move forwarded solve test to dockerfile package
